+++ HISTORY.rst
@@ -6,7 +6,7 @@ Release History
 0.5.2 (2019-07-24)
 ++++++++++++++++++
 
-- sapmonitor create will now accept key vault as well
+- sapmonitor create will also accept keyvault url as credentials
 
 0.5.1 (2019-07-05)
 ++++++++++++++++++